requirements.txt
fastapi==0.115.4
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
pytest==8.3.5
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.37
uvicorn==0.33.0
httpx==0.27.2
pydantic[email]>=2.0.0
python-multipart==0.0.12
python-dotenv==1.0.1